+++ android/src/main/java/com/flutter_webview_plugin/WebviewManager.java
@@ -2,22 +2,33 @@ package com.flutter_webview_plugin;
 
 import android.content.Intent;
 import android.net.Uri;
-import android.util.Log;
 import android.annotation.TargetApi;
 import android.app.Activity;
+import android.content.Context;
 import android.os.Build;
 import android.view.KeyEvent;
 import android.view.View;
 import android.view.ViewGroup;
 import android.webkit.CookieManager;
+import android.webkit.GeolocationPermissions;
 import android.webkit.ValueCallback;
 import android.webkit.WebChromeClient;
+import android.webkit.WebSettings;
 import android.webkit.WebView;
-import android.webkit.WebViewClient;
 import android.widget.FrameLayout;
+import android.provider.MediaStore;
+import androidx.core.content.FileProvider;
+import android.database.Cursor;
+import android.provider.OpenableColumns;
 
+import java.util.List;
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Map;
+import java.io.File;
+import java.util.Date;
+import java.io.IOException;
+import java.text.SimpleDateFormat;
 
 import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
@@ -33,32 +44,45 @@ class WebviewManager {
     private ValueCallback<Uri> mUploadMessage;
     private ValueCallback<Uri[]> mUploadMessageArray;
     private final static int FILECHOOSER_RESULTCODE=1;
+    private Uri fileUri;
+    private Uri videoUri;
+
+    private long getFileSize(Uri fileUri) {
+        Cursor returnCursor = context.getContentResolver().query(fileUri, null, null, null, null);
+        returnCursor.moveToFirst();
+        int sizeIndex = returnCursor.getColumnIndex(OpenableColumns.SIZE);
+        return returnCursor.getLong(sizeIndex);
+    }
 
     @TargetApi(7)
     class ResultHandler {
         public boolean handleResult(int requestCode, int resultCode, Intent intent){
             boolean handled = false;
             if(Build.VERSION.SDK_INT >= 21){
-                Uri[] results = null;
-                // check result
-                if(resultCode == Activity.RESULT_OK){
-                    if(requestCode == FILECHOOSER_RESULTCODE){
-                        if(mUploadMessageArray != null){
-                            String dataString = intent.getDataString();
-                            if(dataString != null){
-                                results = new Uri[]{Uri.parse(dataString)};
-                            }
+                if(requestCode == FILECHOOSER_RESULTCODE){
+                    Uri[] results = null;
+                    if (resultCode == Activity.RESULT_OK) {
+                        if (fileUri != null && getFileSize(fileUri) > 0) {
+                            results = new Uri[] { fileUri };
+                        } else if (videoUri != null && getFileSize(videoUri) > 0) {
+                            results = new Uri[] { videoUri };
+                        } else if (intent != null) {
+                            results = getSelectedFiles(intent);
                         }
-                        handled = true;
                     }
+                    if(mUploadMessageArray != null){
+                        mUploadMessageArray.onReceiveValue(results);
+                        mUploadMessageArray = null;
+                    }
+                    handled = true;
                 }
-                mUploadMessageArray.onReceiveValue(results);
-                mUploadMessageArray = null;
             }else {
                 if (requestCode == FILECHOOSER_RESULTCODE) {
-                    if (null != mUploadMessage) {
-                        Uri result = intent == null || resultCode != RESULT_OK ? null
-                                : intent.getData();
+                	Uri result = null;
+                    if (resultCode == RESULT_OK && intent != null) {
+                        result = intent.getData();
+                    }
+                    if (mUploadMessage != null) {
                         mUploadMessage.onReceiveValue(result);
                         mUploadMessage = null;
                     }
@@ -69,16 +93,39 @@ class WebviewManager {
         }
     }
 
+    private Uri[] getSelectedFiles(Intent data) {
+        // we have one files selected
+        if (data.getData() != null) {
+            String dataString = data.getDataString();
+            if(dataString != null){
+                return new Uri[]{ Uri.parse(dataString) };
+            }
+        }
+        // we have multiple files selected
+        if (data.getClipData() != null) {
+            final int numSelectedFiles = data.getClipData().getItemCount();
+            Uri[] result = new Uri[numSelectedFiles];
+            for (int i = 0; i < numSelectedFiles; i++) {
+                result[i] = data.getClipData().getItemAt(i).getUri();
+            }
+            return result;
+        }
+        return null;
+    }
+
     boolean closed = false;
     WebView webView;
     Activity activity;
+    BrowserClient webViewClient;
     ResultHandler resultHandler;
+    Context context;
 
-    WebviewManager(final Activity activity) {
+    WebviewManager(final Activity activity, final Context context) {
         this.webView = new ObservableWebView(activity);
         this.activity = activity;
+        this.context = context;
         this.resultHandler = new ResultHandler();
-        WebViewClient webViewClient = new BrowserClient();
+        webViewClient = new BrowserClient();
         webView.setOnKeyListener(new View.OnKeyListener() {
             @Override
             public boolean onKey(View v, int keyCode, KeyEvent event) {
@@ -88,7 +135,7 @@ class WebviewManager {
                             if (webView.canGoBack()) {
                                 webView.goBack();
                             } else {
-                                close();
+                                FlutterWebviewPlugin.channel.invokeMethod("onBack", null);
                             }
                             return true;
                     }
@@ -155,20 +202,114 @@ class WebviewManager {
                 }
                 mUploadMessageArray = filePathCallback;
 
-                Intent contentSelectionIntent = new Intent(Intent.ACTION_GET_CONTENT);
-                contentSelectionIntent.addCategory(Intent.CATEGORY_OPENABLE);
-                contentSelectionIntent.setType("*/*");
-                Intent[] intentArray;
-                intentArray = new Intent[0];
+                final String[] acceptTypes = getSafeAcceptedTypes(fileChooserParams);
+                List<Intent> intentList = new ArrayList<Intent>();
+                fileUri = null;
+                videoUri = null;
+                if (acceptsImages(acceptTypes)) {
+                    Intent takePhotoIntent = new Intent(MediaStore.ACTION_IMAGE_CAPTURE);
+                    fileUri = getOutputFilename(MediaStore.ACTION_IMAGE_CAPTURE);
+                    takePhotoIntent.putExtra(MediaStore.EXTRA_OUTPUT, fileUri);
+                    intentList.add(takePhotoIntent);
+                }
+                if (acceptsVideo(acceptTypes)) {
+                    Intent takeVideoIntent = new Intent(MediaStore.ACTION_VIDEO_CAPTURE);
+                    videoUri = getOutputFilename(MediaStore.ACTION_VIDEO_CAPTURE);
+                    takeVideoIntent.putExtra(MediaStore.EXTRA_OUTPUT, videoUri);
+                    intentList.add(takeVideoIntent);
+                }
+                Intent contentSelectionIntent;
+                if (Build.VERSION.SDK_INT >= 21) {
+                    final boolean allowMultiple = fileChooserParams.getMode() == FileChooserParams.MODE_OPEN_MULTIPLE;
+                    contentSelectionIntent = fileChooserParams.createIntent();
+                    contentSelectionIntent.putExtra(Intent.EXTRA_ALLOW_MULTIPLE, allowMultiple);
+                } else {
+                    contentSelectionIntent = new Intent(Intent.ACTION_GET_CONTENT);
+                    contentSelectionIntent.addCategory(Intent.CATEGORY_OPENABLE);
+                    contentSelectionIntent.setType("*/*");
+                }
+                Intent[] intentArray = intentList.toArray(new Intent[intentList.size()]);
 
                 Intent chooserIntent = new Intent(Intent.ACTION_CHOOSER);
                 chooserIntent.putExtra(Intent.EXTRA_INTENT, contentSelectionIntent);
-                chooserIntent.putExtra(Intent.EXTRA_TITLE, "Image Chooser");
                 chooserIntent.putExtra(Intent.EXTRA_INITIAL_INTENTS, intentArray);
                 activity.startActivityForResult(chooserIntent, FILECHOOSER_RESULTCODE);
                 return true;
             }
+
+            @Override
+            public void onProgressChanged(WebView view, int progress) {
+                Map<String, Object> args = new HashMap<>();
+                args.put("progress", progress / 100.0);
+                FlutterWebviewPlugin.channel.invokeMethod("onProgressChanged", args);
+            }
         });
+    }
+
+    private Uri getOutputFilename(String intentType) {
+        String prefix = "";
+        String suffix = "";
+
+        if (intentType == MediaStore.ACTION_IMAGE_CAPTURE) {
+            prefix = "image-";
+            suffix = ".jpg";
+        } else if (intentType == MediaStore.ACTION_VIDEO_CAPTURE) {
+            prefix = "video-";
+            suffix = ".mp4";
+        }
+
+        String packageName = context.getPackageName();
+        File capturedFile = null;
+        try {
+            capturedFile = createCapturedFile(prefix, suffix);
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+        return FileProvider.getUriForFile(context, packageName + ".fileprovider", capturedFile);
+    }
+
+    private File createCapturedFile(String prefix, String suffix) throws IOException {
+        String timeStamp = new SimpleDateFormat("yyyyMMdd_HHmmss").format(new Date());
+        String imageFileName = prefix + "_" + timeStamp;
+        File storageDir = context.getExternalFilesDir(null);
+        return File.createTempFile(imageFileName, suffix, storageDir);
+    }
+
+    private Boolean acceptsImages(String[] types) {
+        return isArrayEmpty(types) || arrayContainsString(types, "image");
+    }
+
+    private Boolean acceptsVideo(String[] types) {
+        return isArrayEmpty(types) || arrayContainsString(types, "video");
+    }
+
+    private Boolean arrayContainsString(String[] array, String pattern) {
+        for (String content : array) {
+            if (content.contains(pattern)) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+    private Boolean isArrayEmpty(String[] arr) {
+        // when our array returned from getAcceptTypes() has no values set from the
+        // webview
+        // i.e. <input type="file" />, without any "accept" attr
+        // will be an array with one empty string element, afaik
+        return arr.length == 0 || (arr.length == 1 && arr[0].length() == 0);
+    }
+
+    private String[] getSafeAcceptedTypes(WebChromeClient.FileChooserParams params) {
+
+        // the getAcceptTypes() is available only in api 21+
+        // for lower level, we ignore it
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
+            return params.getAcceptTypes();
+        }
+
+        final String[] EMPTY = {};
+        return EMPTY;
     }
 
     private void clearCookies() {
@@ -189,18 +330,61 @@ class WebviewManager {
         webView.clearFormData();
     }
 
-    void openUrl(boolean withJavascript, boolean clearCache, boolean hidden, boolean clearCookies, String userAgent, String url, Map<String, String> headers, boolean withZoom, boolean withLocalStorage, boolean scrollBar) {
+    void openUrl(
+            boolean withJavascript,
+            boolean clearCache,
+            boolean hidden,
+            boolean clearCookies,
+            String userAgent,
+            String url,
+            Map<String, String> headers,
+            boolean withZoom,
+            boolean withLocalStorage,
+            boolean scrollBar,
+            boolean supportMultipleWindows,
+            boolean appCacheEnabled,
+            boolean allowFileURLs,
+            boolean useWideViewPort,
+            String invalidUrlRegex,
+            boolean geolocationEnabled
+    ) {
         webView.getSettings().setJavaScriptEnabled(withJavascript);
         webView.getSettings().setBuiltInZoomControls(withZoom);
         webView.getSettings().setSupportZoom(withZoom);
         webView.getSettings().setDomStorageEnabled(withLocalStorage);
+        webView.getSettings().setJavaScriptCanOpenWindowsAutomatically(supportMultipleWindows);
+
+        webView.getSettings().setSupportMultipleWindows(supportMultipleWindows);
+
+        webView.getSettings().setAppCacheEnabled(appCacheEnabled);
+
+        webView.getSettings().setAllowFileAccessFromFileURLs(allowFileURLs);
+        webView.getSettings().setAllowUniversalAccessFromFileURLs(allowFileURLs);
+
+        webView.getSettings().setUseWideViewPort(useWideViewPort);
+
+        webViewClient.updateInvalidUrlRegex(invalidUrlRegex);
+
+        if (geolocationEnabled) {
+            webView.getSettings().setGeolocationEnabled(true);
+            webView.setWebChromeClient(new WebChromeClient() {
+                @Override
+                public void onGeolocationPermissionsShowPrompt(String origin, GeolocationPermissions.Callback callback) {
+                    callback.invoke(origin, true, false);
+                }
+            });
+        }
+
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
+            webView.getSettings().setMixedContentMode(WebSettings.MIXED_CONTENT_COMPATIBILITY_MODE);
+        }
 
         if (clearCache) {
             clearCache();
         }
 
         if (hidden) {
-            webView.setVisibility(View.INVISIBLE);
+            webView.setVisibility(View.GONE);
         }
 
         if (clearCookies) {
@@ -210,7 +394,7 @@ class WebviewManager {
         if (userAgent != null) {
             webView.getSettings().setUserAgentString(userAgent);
         }
-      
+
         if(!scrollBar){
             webView.setVerticalScrollBarEnabled(false);
         }
@@ -220,6 +404,10 @@ class WebviewManager {
         } else {
             webView.loadUrl(url);
         }
+    }
+
+    void reloadUrl(String url) {
+        webView.loadUrl(url);
     }
 
     void close(MethodCall call, MethodChannel.Result result) {
@@ -251,7 +439,7 @@ class WebviewManager {
             }
         });
     }
-    /** 
+    /**
     * Reloads the Webview.
     */
     void reload(MethodCall call, MethodChannel.Result result) {
@@ -259,7 +447,7 @@ class WebviewManager {
             webView.reload();
         }
     }
-    /** 
+    /**
     * Navigates back on the Webview.
     */
     void back(MethodCall call, MethodChannel.Result result) {
@@ -267,7 +455,7 @@ class WebviewManager {
             webView.goBack();
         }
     }
-    /** 
+    /**
     * Navigates forward on the Webview.
     */
     void forward(MethodCall call, MethodChannel.Result result) {
@@ -279,13 +467,13 @@ class WebviewManager {
     void resize(FrameLayout.LayoutParams params) {
         webView.setLayoutParams(params);
     }
-    /** 
+    /**
     * Checks if going back on the Webview is possible.
     */
     boolean canGoBack() {
         return webView.canGoBack();
     }
-    /** 
+    /**
     * Checks if going forward on the Webview is possible.
     */
     boolean canGoForward() {
@@ -293,7 +481,7 @@ class WebviewManager {
     }
     void hide(MethodCall call, MethodChannel.Result result) {
         if (webView != null) {
-            webView.setVisibility(View.INVISIBLE);
+            webView.setVisibility(View.GONE);
         }
     }
     void show(MethodCall call, MethodChannel.Result result) {